# expression: implement vectorized builtinFormatSig