--- a/expression/builtin_string_vec_test.go
+++ b/expression/builtin_string_vec_test.go
@@ -481,6 +481,24 @@ var vecBuiltinStringCases = map[string][]vecExprBenchCase{
 			),
 		}},
 	},
+	ast.Format: {
+		{retEvalType: types.ETString, childrenTypes: []types.EvalType{types.ETDecimal, types.ETInt}, geners: []dataGenerator{
+			newRangeDecimalGener(-10000, 10000, 0),
+			newRangeInt64Gener(-10, 40),
+		}},
+		{retEvalType: types.ETString, childrenTypes: []types.EvalType{types.ETReal, types.ETInt}, geners: []dataGenerator{
+			newRangeRealGener(-10000, 10000, 0),
+			newRangeInt64Gener(-10, 40),
+		}},
+		{retEvalType: types.ETString, childrenTypes: []types.EvalType{types.ETDecimal, types.ETInt}, geners: []dataGenerator{
+			newRangeDecimalGener(-10000, 10000, 1),
+			newRangeInt64Gener(-10, 40),
+		}},
+		{retEvalType: types.ETString, childrenTypes: []types.EvalType{types.ETReal, types.ETInt}, geners: []dataGenerator{
+			newRangeRealGener(-10000, 10000, 1),
+			newRangeInt64Gener(-10, 40),
+		}},
+	},
 }
 
 func (s *testVectorizeSuite1) TestVectorizedBuiltinStringEvalOneVec(c *C) {