<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

PCP #12106
### What problem does this PR solve? <!--add issue link with summary if exists-->
Implement builtinFormatSig vectorized function
#12106

### What is changed and how it works?
- Implement builtinFormatSig vectorized function
- Add 5 test cases for vectorized builtinFormatSig:
    - decimal `x` with int64 `d`
    - real `x` with int64 d
    - decimal null `x` with int64 `d`
    - real null `x` with int64 `d`
    - real string `x` with int64 string `d`

```
$ go test -v -benchmem -bench=BenchmarkVectorizedBuiltinStringFunc -run=BenchmarkVectorizedBuiltinStringFunc -args builtinFormatSig
goos: darwin
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-VecBuiltinFunc-8              1392            834639 ns/op          297394 B/op       8746 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-NonVecBuiltinFunc-8           1318            882052 ns/op          297392 B/op       8746 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-VecBuiltinFunc#01-8           1216            955453 ns/op          298992 B/op       8773 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-NonVecBuiltinFunc#01-8        1171           1051933 ns/op          298992 B/op       8773 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-VecBuiltinFunc#02-8         183423              6299 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-NonVecBuiltinFunc#02-8       37773             32145 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-VecBuiltinFunc#03-8         254294              4615 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-NonVecBuiltinFunc#03-8       38475             31142 ns/op               0 B/op          0 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-VecBuiltinFunc#04-8           1339            878812 ns/op          214512 B/op       7600 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinFormatSig-NonVecBuiltinFunc#04-8        1147           1023566 ns/op          214512 B/op       7600 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      13.319s
```


### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test